+++ CHANGELOG.md
@@ -35,9 +35,14 @@ Section Order:
 ### Security
 -->
 
+### Added
+
+- Bootstrap tooltips
+
 ### Changed
 
 - "View all/active" button now is only shown when in the SRP links view
+- Chinese translation updated
 - French translation updated
 
 ## \[2.2.0\] - 2024-06-18